fix: fix support for Azure AD provider
When using https://github.com/TheNetworg/oauth2-azure you can't log in using your AzureAD tenant, because Azure doesn't return the email as an "email" property but as "upn" property. If we find UPN set in the response then we're assigning it to the $email and we set $email_verified to True.